ai-sql-agent-backend/pyproject.toml
[tool.poetry]
name = "db-chatapp"
version = "0.1.0"
description = "AI-powered application that allows users to interact with csv/excel/database"
authors = ["Jayanth MKV <edithultron123@gmail.com>"]
license = "Apache License"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
langchain = "^0.2.12"
langchain-groq = "^0.1.9"
pandas = "^2.2.2"
langchain-community = "^0.2.11"
python-dotenv = "^1.0.1"
langgraph = "^0.2.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
langchainhub = "^0.1.20"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
